requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
asyncmy==0.2.9
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-spinner==0.1.10
Crawl4AI==0.3.73
croniter==1.4.1
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.3
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastapi-utilities==0.3.0
fastapi-utils==0.7.0
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
litellm==1.52.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
openai==1.54.1
orjson==3.10.11
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
playwright==1.47.0
playwright-stealth==1.0.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
PyMySQL==1.1.1
pypika-tortoise==0.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.2
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
tornado==6.4.1
tortoise-orm==0.21.7
tqdm==4.66.6
traitlets==5.14.3
typer==0.13.0
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
wheel==0.44.0
yarg==0.1.9
yarl==1.17.1
zipp==3.20.2
